updated worksamples routes with new controller workSampleSetUp
